Split batch into two files for old and new
