+++ 2017-10/ClientAPIs/adam/__init__.py
@@ -4,10 +4,10 @@
 
 from adam.auth import Auth
 from adam.batch import Batch
-from adam.batch import Batch2
-from adam.batch import PropagationParams
-from adam.batch import OpmParams
-from adam.batch import Batches
+from adam.batch2 import Batch2
+from adam.batch2 import PropagationParams
+from adam.batch2 import OpmParams
+from adam.batch2 import Batches
 from adam.batch_run_manager import BatchRunManager
 from adam.group import Groups
 from adam.project import Projects
